Improved: Added barcode identifier on the store view's settings page(#584)
