+++ src/views/StorePermissions.vue
@@ -23,24 +23,7 @@
             </ion-toggle>
           </ion-item>
         </ion-card>
-        <ion-card>
-          <ion-card-header>
-            <ion-card-title>
-              {{ translate('Force scan') }}
-            </ion-card-title>
-          </ion-card-header>
-          <ion-card-content v-html="barcodeContentMessage"></ion-card-content>
-          <ion-item lines="none">
-            <ion-toggle :checked="productStoreSettings['forceScan']" @click.prevent="updateProductStoreSetting($event, 'forceScan')">
-              {{ translate("Require barcode scanning") }}
-            </ion-toggle>
-          </ion-item>
-          <ion-item lines="none">
-            <ion-select :label="translate('Barcode Identifier')" interface="popover" :placeholder="translate('Select')" :value="productStoreSettings['barcodeIdentificationPref']" @ionChange="setBarcodeIdentificationPref($event.detail.value)">
-              <ion-select-option v-for="identification in productIdentifications" :key="identification" :value="identification" >{{ identification }}</ion-select-option>
-            </ion-select>
-          </ion-item>
-        </ion-card>
+        <ForceScanCard />
       </div>
     </ion-content> 
   </ion-page>
@@ -56,8 +39,6 @@ import {
   IonHeader,
   IonItem,
   IonPage,
-  IonSelect,
-  IonSelectOption,
   IonTitle,
   IonToggle,
   IonToolbar,
@@ -66,23 +47,17 @@ import {
 import { translate } from '@/i18n'
 import store from "@/store";
 import { computed } from "vue";
+import ForceScanCard from "@/components/ForceScanCard.vue";
 
-const barcodeContentMessage = translate("Require inventory to be scanned when counting instead of manually entering values. If the identifier is not found, the scan will default to using the internal name.", { space: '<br /><br />' })
 const productStoreSettings = computed(() => store.getters["user/getProductStoreSettings"])
-const productIdentifications = computed(() => store.getters["user/getGoodIdentificationTypes"])
 
 onIonViewWillEnter(async () => {
   await store.dispatch("user/getProductStoreSetting")
-  await store.dispatch("user/fetchGoodIdentificationTypes")
 })
 
 function updateProductStoreSetting(event: any, key: string) {
   event.stopImmediatePropagation();
   store.dispatch("user/setProductStoreSetting", { key, value: !productStoreSettings.value[key] })
-}
-
-function setBarcodeIdentificationPref(value: any) {
-  store.dispatch("user/setProductStoreSetting", { key: "barcodeIdentificationPref", value })
 }
 </script>
